+++ ticketguru/src/main/java/copypaste/ticketguru/web/rest/EventRestController.java
@@ -1,7 +1,13 @@
 package copypaste.ticketguru.web.rest;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,22 +23,30 @@ import org.springframework.web.bind.annotation.RestController;
 import copypaste.ticketguru.domain.Event;
 import copypaste.ticketguru.domain.EventRepository;
 import copypaste.ticketguru.domain.EventRequest;
+import copypaste.ticketguru.domain.TicketType;
+import copypaste.ticketguru.domain.TicketTypeRepository;
+import copypaste.ticketguru.domain.TicketTypeRequest;
 
 @RestController
 public class EventRestController {
     @Autowired
-    EventRepository repository;
+    EventRepository eventRepository;
+    // Looks evil, should probably have its own file, but don't know how to make it nice in that case
+    @Autowired
+    TicketTypeRepository ticketTypeRepository;
+
+    private static final Set<String> ALLOWED_TICKET_TYPES = Set.of("Aikuinen", "Lapsi", "Eläkeläinen", "Opiskelija", "Varusmies", "VIP");
 
     // hae kaikki tapahtumat
 	@GetMapping(value = "/api/events")
 	public List<Event> getAllEvents() {
-		return (List<Event>) repository.findAll();
+		return (List<Event>) eventRepository.findAll();
 	}
 
 	// hae yksi tapahtuma ID:llä
 	@GetMapping(value = "/api/events/{id}")
 	public ResponseEntity<Event> getEventById(@PathVariable Long id) {
-		Optional<Event> eventOpt = repository.findById(id);
+		Optional<Event> eventOpt = eventRepository.findById(id);
 		if (!eventOpt.isPresent()) {
 			return ResponseEntity.notFound().build();
 		}
@@ -42,7 +56,7 @@ public class EventRestController {
 	// hae tapahtumat kaupungin mukaan
 	@GetMapping(value = "/api/events/byName")
 	public ResponseEntity<List<Event>> findEventsByName(@RequestParam("name") String name) {
-		List<Event> events = repository.findByNameContainingIgnoreCase(name);
+		List<Event> events = eventRepository.findByNameContainingIgnoreCase(name);
 		if (events.isEmpty()) {
 			return ResponseEntity.noContent().build();
 		}
@@ -52,7 +66,7 @@ public class EventRestController {
 	// hae tapahtumat tapahtuman nimen mukaan
 	@GetMapping(value = "/api/events/byCity")
 	public ResponseEntity<List<Event>> findEventsByCity(@RequestParam("city") String city) {
-		List<Event> events = repository.findByCityIgnoreCase(city);
+		List<Event> events = eventRepository.findByCityIgnoreCase(city);
 		if (events.isEmpty()) {
 			return ResponseEntity.noContent().build();
 		}
@@ -62,29 +76,97 @@ public class EventRestController {
     // Lisää tapahtuma
 	@PostMapping(value = "/api/events")
 	public ResponseEntity<Event> createEvent(@RequestBody Event newEvent) {
-		Event savedEvent = repository.save(newEvent);
-		return ResponseEntity.status(HttpStatus.CREATED).body(savedEvent);
+		return ResponseEntity
+            .status(HttpStatus.CREATED)
+            .body(eventRepository.save(newEvent));
 	}
 
+    // Tapahtuman päivitys
     @PutMapping(value = "/api/events/{id}")
 	public ResponseEntity<?> updateEvent(@PathVariable Long id, @RequestBody EventRequest eventRequest) {
-		return repository.findById(id).map(event -> {
+		return eventRepository.findById(id).map(event -> {
 			event.setName(eventRequest.getName());
 			event.setDate(eventRequest.getDate());
 			event.setPlace(eventRequest.getPlace());
 			event.setCity(eventRequest.getCity());
 			event.setTicketCount(eventRequest.getTicketCount());
 
-			Event updatedEvent = repository.save(event);
+			Event updatedEvent = eventRepository.save(event);
 			return ResponseEntity.ok(updatedEvent);
 		}).orElse(ResponseEntity.notFound().build());
 	}
 
+    // Tapahtuman poisto
     @DeleteMapping("/api/delete/{id}")
 	public ResponseEntity<?> deleteEvent(@PathVariable long id) {
-        return repository.findById(id).map(event -> {
-            repository.delete(event);
+        return eventRepository.findById(id).map(event -> {
+            eventRepository.delete(event);
             return ResponseEntity.noContent().build();
+        }).orElse(ResponseEntity.notFound().build());
+	}
+
+    // Lisää lipputyypit tapahtumaan
+	@PostMapping("/api/events/{eventId}/tickettypes")
+	public ResponseEntity<?> createTicketTypesForEvent(@PathVariable Long eventId, @RequestBody List<TicketTypeRequest> ticketTypeRequests) {
+		// Tarkastetaan löytyykö tapahtuma id:n perusteella
+        return eventRepository.findById(eventId).map(event -> {
+            List<TicketType> ticketTypes = new ArrayList<>();
+            for (TicketTypeRequest ttRequest : ticketTypeRequests) {
+                if (!ALLOWED_TICKET_TYPES.contains(ttRequest.getName())) {
+                    return ResponseEntity.badRequest().body("Ticket type " + ttRequest.getName() + " is not allowed.");
+                }
+                TicketType newTicketType = new TicketType(ttRequest.getName(), ttRequest.getPrice(), event);
+                ticketTypes.add(newTicketType);
+            }
+
+            List<TicketType> savedTicketTypes = (List<TicketType>) ticketTypeRepository.saveAll(ticketTypes);
+
+            // Yksinkertaistetaan JSON-vastausta
+            List<Map<String, Object>> response = savedTicketTypes.stream().map(tt -> {
+                Map<String, Object> ticketTypeResponse = new HashMap<>();
+                ticketTypeResponse.put("id", tt.getId());
+                ticketTypeResponse.put("name", tt.getName());
+                ticketTypeResponse.put("price", tt.getPrice());
+                ticketTypeResponse.put("event", tt.getEvent().getId()); // sisällytä vain ID
+                return ticketTypeResponse;
+            }).collect(Collectors.toList());
+
+            return ResponseEntity.status(HttpStatus.CREATED).body(response);
+        }).orElse(ResponseEntity.notFound().build());
+	}
+
+    // Päivität tapahtuman lipputyyppejä
+	@PutMapping("/api/events/{eventId}/tickettypes")
+	public ResponseEntity<?> updateTicketTypesForEvent(@PathVariable Long eventId, @RequestBody List<TicketTypeRequest> ticketTypeRequests) {
+        // Tarkastetaan löytyykö tapahtuma id:n perusteella
+        return eventRepository.findById(eventId).map(event -> {
+            List<TicketType> existingTicketTypes = ticketTypeRepository.findByEvent(event);
+            // Poistetaan aikaisemmat lipputyypit
+            ticketTypeRepository.deleteAll(existingTicketTypes);
+
+            // Tarkastetaan lipputyyppien nimet, luodaan uudet lipputyypit ja määritetään mihin tapahtumaan ne kuuluvat
+            List<TicketType> ticketTypes = new ArrayList<>();
+            for (TicketTypeRequest ttRequest : ticketTypeRequests) {
+                if (!ALLOWED_TICKET_TYPES.contains(ttRequest.getName())) {
+                    return ResponseEntity.badRequest().body("Ticket type " + ttRequest.getName() + " is not allowed.");
+                }
+                TicketType newTicketType = new TicketType(ttRequest.getName(), ttRequest.getPrice(), event);
+                ticketTypes.add(newTicketType);
+            }
+
+            List<TicketType> savedTicketTypes = (List<TicketType>) ticketTypeRepository.saveAll(ticketTypes);
+
+            // Yksinkertaistetaan JSON-vastausta
+            List<Map<String, Object>> response = savedTicketTypes.stream().map(tt -> {
+                Map<String, Object> ticketTypeResponse = new HashMap<>();
+                ticketTypeResponse.put("id", tt.getId());
+                ticketTypeResponse.put("name", tt.getName());
+                ticketTypeResponse.put("price", tt.getPrice());
+                ticketTypeResponse.put("event", tt.getEvent().getId()); // Sisällytetään vastaukseen vain tapahtuman id
+                return ticketTypeResponse;
+            }).collect(Collectors.toList());
+
+            return ResponseEntity.ok(response);
         }).orElse(ResponseEntity.notFound().build());
 	}
 }
