+++ ticketguru/src/main/java/copypaste/ticketguru/web/rest/PurchaseRestController.java
@@ -1,0 +1,5 @@
+package copypaste.ticketguru.web.rest;
+
+public class PurchaseRestController {
+    
+}
